Making all messages SerializationAware.
